Añadir componentes de TaskCard y mejorar la gestión de documentos en varios formularios
