+++ src/app/[locale]/(home)/agents-management/recruitment/components/AddDocuments.jsx
@@ -29,13 +29,13 @@ const AddDocuments = ({ contactId }) => {
     },
   ];
 
-  const handleAddDocument = (document) => {
+  const handleAddDocument = (documentToAdd) => {
     setAddFileProps({
       ...addFileProps,
       isOpen: true,
-      documentType: document.type,
-      title: t("common:add-document", { document: document.name }),
-      accept: document.accept,
+      documentType: documentToAdd?.type,
+      title: t("common:add-document", { document: documentToAdd?.name }),
+      accept: documentToAdd?.accept,
     });
   };
