+++ src/app/[locale]/(home)/home/components/CompleteModalTask.jsx
@@ -1,3 +1,4 @@
+"use client";
 import Button from "@/src/components/form/Button";
 import {
   Description,
