+++ src/app/[locale]/(home)/home/components/ContactList.jsx
@@ -1,3 +1,4 @@
+"use client";
 import clsx from "clsx";
 import Link from "next/link";
 import { Fragment } from "react";
