+++ src/app/[locale]/(home)/sales/crm/leads/components/AddDocuments.jsx
@@ -23,13 +23,13 @@ const AddDocuments = ({ leadId }) => {
     { name: t("leads:add:policy"), type: "poliza", accept: ".pdf" },
   ];
 
-  const handleAddDocument = (document) => {
+  const handleAddDocument = (documentToAdd) => {
     setAddFileProps({
       ...addFileProps,
       isOpen: true,
-      documentType: document.type,
-      title: t("common:add-document", { document: document.name }),
-      accept: document.accept,
+      documentType: documentToAdd?.type,
+      title: t("common:add-document", { document: documentToAdd?.name }),
+      accept: documentToAdd?.accept,
     });
   };
